saving work thus far
