+++ public/account.html
@@ -16,7 +16,12 @@
     type="text/javascript"
     src="js/header.js"
     ></script>
+
+    <script src="https://code.jquery.com/jquery-3.5.1.js" integrity="sha256-QWo7LDvxbWT2tbbQ97B53yJnYU3WhH/C8ycbRAkjPDc=" crossorigin="anonymous"></script>
+    <script src="./js/broker.js"></script>
+    
     <h2>Welcome To Your Account!</h2>
+    <h2 id="pi"></h2>
   
     
     <div class="ui placeholder">
@@ -31,9 +36,9 @@
           </div>
       </div>
       <div>
-          <div class="uk-card uk-card-default uk-card-hover uk-card-body">
-              <h3 class="uk-card-title">Default</h3>
-              <p>Lorem ipsum dolor sit amet, consectetur adipisicing elit.</p>
+          <div id="userPlaylists" class="uk-card uk-card-default uk-card-hover uk-card-body">
+              <h3 class="uk-card-title">PLaylists</h3>
+              
           </div>
       </div>
       <div>
