+++ public/join.html
@@ -5,40 +5,65 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/uikit@3.5.16/dist/css/uikit.min.css" />
-    <link rel="stylesheet" href="https://unpkg.com/purecss@2.0.3/build/pure-min.css" integrity="sha384-4ZPLezkTZTsojWFhpdFembdzFudphhoOzIunR1wH6g1WQDzCAiPvDyitaK67mp0+" crossorigin="anonymous">
     <link href="https://fonts.googleapis.com/css2?family=Audiowide&display=swap" rel="stylesheet">
     <link rel="stylesheet" href="css/styles.css">
     <title>5Wire</title>
 </head>
 
 <body>
-    <script language="javascript" type="text/javascript" src="js/header.js"></script>
-
-
-
-    <form class="pure-form pure-form-aligned">
-        <fieldset class="join-form">
-            <div class="pure-control-group">
-                <label for="aligned-name">Username</label>
-                <input type="text" id="aligned-name" placeholder="Username" />
-                <span class="pure-form-message-inline">This is a required field.</span>
+    <body>
+        <script
+          language="javascript"
+          type="text/javascript"
+          src="js/header.js"
+        ></script>
+        <script src="https://code.jquery.com/jquery-3.5.1.js" integrity="sha256-QWo7LDvxbWT2tbbQ97B53yJnYU3WhH/C8ycbRAkjPDc=" crossorigin="anonymous"></script>
+        <script src="./js/login.js"></script>
+        <section class="form">
+          <h1>Create New Account</h1>
+          <form id="loginForm" class="uk-form-horizontal uk-margin-large">
+            <div class="uk-margin">
+              <label class="uk-form-label" for="form-horizontal-text"
+                >Username</label
+              >
+              <div class="uk-form-controls">
+                <input
+                  class="uk-input"
+                  id="userName"
+                  type="text"
+                  placeholder="Username"
+                  name="un"
+                />
+              </div>
+              <div class="uk-margin">
+                <label class="uk-form-label" for="form-horizontal-text"
+                  >Password</label
+                >
+                <div class="uk-form-controls">
+                  <input
+                    class="uk-input"
+                    id="passWord"
+                    type="password"
+                    placeholder="Password"
+                    name="pw"
+                  />
+                </div>
+              </div>
+              <div>
+                <a action="submit" id="login" class="uk-button uk-button-default" href="#">I already have an account</a>
+                <a action="submit" id="join" class="uk-button uk-button-default" href="#">Join</a>
+              </div>
             </div>
-            <div class="pure-control-group">
-                <label for="aligned-password">Password</label>
-                <input type="password" id="aligned-password" placeholder="Password" />
-            </div>
-            <div class="pure-controls">
-                <label for="aligned-cb" class="pure-checkbox">
-                    <input type="checkbox" id="aligned-cb" /> I&#x27;ve read the terms and conditions</label>
-                <button type="submit" class="pure-button pure-button-primary">Submit</button>
-            </div>
-        </fieldset>
-    </form>
-
-
-    <script language="javascript" type="text/javascript" src="js/footer.js"></script>
-    <script src="https://cdn.jsdelivr.net/npm/uikit@3.5.16/dist/js/uikit.min.js"></script>
-    <script src="https://cdn.jsdelivr.net/npm/uikit@3.5.16/dist/js/uikit-icons.min.js"></script>
+          </form>
+        </section>
+    
+        <script
+          language="javascript"
+          type="text/javascript"
+          src="js/footer.js"
+        ></script>
+        <script src="https://cdn.jsdelivr.net/npm/uikit@3.5.16/dist/js/uikit.min.js"></script>
+        <script src="https://cdn.jsdelivr.net/npm/uikit@3.5.16/dist/js/uikit-icons.min.js"></script>
 </body>
 
 </html>
