+++ public/js/header.js
@@ -7,7 +7,7 @@
 //something like if login=true then display navagation A else display navagtion B
 
 
-document.write('<center>  <nav class="uk-navbar-container uk-margin" uk-navbar><div class="uk-navbar-center"><div class="uk-navbar-center-left"><div><ul class="uk-navbar-nav"><li><a href="#">Home</a></li><li><a href="#">Account</a></li><li><a href="#">Create Playlist</a></li></ul></div></div><a class="uk-navbar-item uk-logo" href="index.html"><img src="images/5wire_logo.png" width="100px"height="auto"></a><div class="uk-navbar-center-right"><div><ul class="uk-navbar-nav"><li><a href="#">Log Out</a></li></ul></div></div></div></nav></center>')
+document.write('<center>  <nav class="uk-navbar-container uk-margin" uk-navbar><div class="uk-navbar-center"><div class="uk-navbar-center-left"><div><ul class="uk-navbar-nav"><li><a href="/">Home</a></li><li><a href="/account">Account</a></li><li><a href="/create">Create Playlist</a></li></ul></div></div><a class="uk-navbar-item uk-logo" href="index.html"><img src="images/5wire_logo.png" width="100px"height="auto"></a><div class="uk-navbar-center-right"><div><ul class="uk-navbar-nav"><li><a href="#">Log Out</a></li></ul></div></div></div></nav></center>')
 
 //instructions for adding this to the html
 //<BODY>
