+++ public/playlist.html
@@ -6,20 +6,64 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/uikit@3.5.16/dist/css/uikit.min.css" />
     <link href="https://fonts.googleapis.com/css2?family=Audiowide&display=swap" rel="stylesheet">
-    <link rel="stylesheet" href="css/styles.css">
+    <link rel="stylesheet" href="../css/styles.css">
     <title>5Wire</title>
 </head>
 
 <body>
-    <script language="javascript" type="text/javascript" src="js/header.js"></script>
+    <center>  <nav class="uk-navbar-container uk-margin" uk-navbar><div class="uk-navbar-center"><div class="uk-navbar-center-left"><div><ul class="uk-navbar-nav"><li><a href="/">Home</a></li><li><a href="/account">Account</a></li><li><a href="/create">Create Playlist</a></li></ul></div></div><a class="uk-navbar-item uk-logo" href="index.html"><img src="../images/5wire_logo.png" width="100px"height="auto"></a><div class="uk-navbar-center-right"><div><ul class="uk-navbar-nav"><li><a href="#">Log Out</a></li></ul></div></div></div></nav></center>
+
+    <script src="https://code.jquery.com/jquery-3.5.1.js" integrity="sha256-QWo7LDvxbWT2tbbQ97B53yJnYU3WhH/C8ycbRAkjPDc=" crossorigin="anonymous"></script>
+    <script src="../js/playlist.js"></script>
 
     <header id="banner">
         <div class="uk-container">
-            <h1>Playlist Name</h1>
+            <h1 id="plTitle"></h1>
+            <h2 id="plCat"></h2>
+            <p id="plDesc"></p>
             <button id="play" class="uk-button uk-button-default" href="#">Play</button>
         </div>
     </header>
-    <section class="uk-container">
+    <div class="uk-grid">
+    <div class="uk-column uk-width-1-3@m">
+        <ol id="plSongs"></ol>
+    </div>
+    <div class="uk-column uk-width-2-3@m">
+    <div id="spotifyDiv">
+        <form id="spotifyForm" class="uk-form-horizontal uk-margin-large">
+            <div class="uk-margin">
+                <div class="uk-form-controls">
+                    <select class="uk-select" id="type" placeholder="category">
+                        <option name="track">Track</option>
+                        <!-- <option name="artist">Artist</option>
+                        <option name="album">Album</option> -->
+                    </select>
+                  </div>
+              <label class="uk-form-label" for="form-horizontal-text"
+                >Song</label
+              >
+              <div class="uk-form-controls">
+                <input
+                  class="uk-input"
+                  id="search"
+                  type="text"
+                  placeholder="Type in a Song"
+                  name="un"
+                />
+              </div>
+              <div>
+                <a action="submit" id="searchSpotify" class="uk-button uk-button-default" href="#">Search</a>
+              </div>
+            </div>
+          </form>
+    
+        <div id="spotResults">
+            
+        </div>
+    </div>
+</div>
+</div>
+    <!-- <section class="uk-container">
 <br><br>
         <div class="uk-column-1-3">
             <p>Song 1</p>
@@ -41,11 +85,11 @@
         </div>
         <hr>
         <br><br>
-    </section>
+    </section> -->
 
 
 
-    <script language="javascript" type="text/javascript" src="js/footer.js"></script>
+    <script language="javascript" type="text/javascript" src="../js/footer.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/uikit@3.5.16/dist/js/uikit.min.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/uikit@3.5.16/dist/js/uikit-icons.min.js"></script>
 </body>
